+++ aedile-core/src/test/kotlin/com/sksamuel/aedile/core/EvictionTest.kt
@@ -4,6 +4,7 @@ import com.github.benmanes.caffeine.cache.RemovalCause
 import io.kotest.assertions.timing.eventually
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import kotlinx.coroutines.delay
 import kotlin.time.Duration.Companion.seconds
 
 class EvictionTest : FunSpec() {
@@ -13,6 +14,23 @@ class EvictionTest : FunSpec() {
          val cache = caffeineBuilder<String, String> {
             maximumSize = 1
             evictionListener = { _, _, removalCause -> cause = removalCause }
+         }.build()
+         repeat(2) { k ->
+            cache.put("$k") { "bar" }
+         }
+         eventually(5.seconds) {
+            cause shouldBe RemovalCause.SIZE
+         }
+      }
+
+      test("cache should support suspendable eviction functions") {
+         var cause: RemovalCause? = null
+         val cache = caffeineBuilder<String, String> {
+            maximumSize = 1
+            evictionListener = { _, _, removalCause ->
+               delay(1)
+               cause = removalCause
+            }
          }.build()
          repeat(2) { k ->
             cache.put("$k") { "bar" }